+++ client/src/components/Mypage/Mypage.jsx
@@ -1,10 +1,12 @@
-﻿import React, { useState, useRef } from 'react';
-import { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
+﻿import React, { useState, useRef, useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import * as S from "./Mypage.style";
+
+import Info from './info/Info';
 import Post from './post/Post';
 import Comment from './comment/Comment';
 import Bookmark from './bookmark/Bookmark';
+
 import img from "assets/images/mypage_back_Img1.png";
 import profileImg from "assets/images/profile_img.png";
 
@@ -12,10 +14,8 @@ import { mypageBookmarkMiddleware } from 'store/modules/mypageBookmark';
 import { mypageCommentMiddleware } from 'store/modules/mypageComment';
 import { mypagePostMiddleware } from 'store/modules/mypagePost';
 
-import { dbService, storageService } from 'firebase.js';
 import firebase from 'firebase';
 
-import { v4 as uuidv4 } from "uuid";
 
 
 const Mypage = ({user}) => {
@@ -24,22 +24,23 @@ const Mypage = ({user}) => {
   const [post, setPost] = useState(false);
   const [comment, setComment] = useState(false);
   const [bookmark, setBookmark] = useState(false);
-  const [attachment, setAttachment] = useState([]);
+
+  // const [change, setChange] = useState(false);
 
   const [uid, setUid] = useState('');
-
-  const input = useRef();
 
   const auth = firebase.auth();
 
   const dispatch = useDispatch();
+
+  const userDB = useSelector(state => state.userLike.data);
 
   const onClose = () => {
     setInfo(false);
     setPost(false);
     setComment(false);
     setBookmark(false);
-  }
+  };
 
   const onInfo = () => {
     if(info){
@@ -57,7 +58,7 @@ const Mypage = ({user}) => {
     }
     onClose();
     setPost(true);
-  }
+  };
 
   const onComment = () => {
     if(comment){
@@ -66,7 +67,7 @@ const Mypage = ({user}) => {
     }
     onClose();
     setComment(true);
-  }
+  };
 
   const onBookmark = () => {
     if(bookmark){
@@ -75,68 +76,8 @@ const Mypage = ({user}) => {
     }
     onClose();
     setBookmark(true);
-  }
-
-  const onUsername = (value) => {
-    dbService.collection('users').doc(user.user_id).update({
-      user_name: value,      
-    });
-  }
-
-  const onUserImage = (value) => {
-    dbService.collection('users').doc(user.user_id).update({
-      user_image: value,      
-    });
-  }
-
-  const onFileChange = (e) => {
-    const { files } = e.target;
-    let file;
-    let fileURLs = [];
-
-    file = files[0];
-    let reader = new FileReader();
-    reader.onload = () => {
-      fileURLs[0] = reader.result;
-      setAttachment([...fileURLs]);
-    };
-    reader.readAsDataURL(file);
   };
 
-  const onEnter = (e) => {
-    if(e.key != 'Enter'){
-      return;
-    }
-    if(e.key === 'Enter'){
-      e.preventDefault();
-      onSubmitBtn();
-    }
-  }
-
-  const onSubmitBtn = () => {
-    onUsername(input.current.value);
-    input.current.value = '';
-  }
-
-  const onSubmit = async(e) => {
-    e.preventDefault();
-    let attachmentUrl = [];
-    if (attachment) {
-      const attachmentRef = storageService
-        .ref()
-        .child(`${uid.uid}/${uuidv4()}`);
-      const response = await attachmentRef.putString(
-        attachment[0],
-        "data_url"
-      );
-      attachmentUrl.push(await response.ref.getDownloadURL());
-    };
-
-    await dbService.collection('users').doc(user.user_id).update({
-      user_image: attachmentUrl,
-    });
-    setAttachment([]);
-  };
 
   useEffect(()=>{
     dispatch(mypagePostMiddleware(user.user_id));
@@ -162,13 +103,21 @@ const Mypage = ({user}) => {
     });
   },[]);
 
+  // useEffect(()=>{
+  //   // setChange(false);
+  //   console.log(change);
+  // },[change]);
+
 
   return (
     <>
       <S.Container>
         <S.Contents>
           <S.BackImage>
-            <img src={user.user_image} alt="배경사진" />
+            {/* {change ?
+            <img src={userDB.user_image} alt="배경사진" /> :
+            ((userDB.user_image !== user.user_image) && <img src={user.user_image} alt="배경사진" />)} */}
+            <img src={userDB.user_image} alt="배경사진" />
           </S.BackImage>
           <S.ListArea>
             <p onClick={onInfo}>
@@ -194,30 +143,8 @@ const Mypage = ({user}) => {
         {info &&
         <S.Content>
           <li>내 정보</li>
-          <li>{user.user_id}</li>
-          {user.user_name ? <li>{user.user_name}</li> : <li onClick={onUsername}>닉네임 설정할래?</li>}
-          {user.user_image ? <li><img src={user.user_image} alt="프로필 사진"></img></li> : <li onClick={onUserImage}>사진 추가할래?</li>}
           <li>
-            <input ref={input} type="text" onKeyPress={e=>onEnter(e)}/>
-            <button onClick={onSubmitBtn}>제출</button>
-          </li>
-          <li>
-            <form onSubmit={onSubmit}>
-              <input
-                accept="image/*"
-                type="file"
-                onChange={onFileChange}
-                name="fileNames[]"
-              />
-              {attachment && (
-                <div>
-                  {attachment.map((atta, i) => (
-                    <img key={i} src={atta} width="50px" height="50px" alt="프로필 사진"/>
-                  ))}
-                </div>
-              )}
-              <button type="submit">제출</button>
-            </form>
+            <Info uid={uid} user={user} userDB={userDB} />
           </li>
         </S.Content>}
         {post &&
